move db test to separate file
